progress on restricted notes
