+++ spec/javascripts/student_profile/restricted_notes_page_container_spec.jsx
@@ -2,10 +2,10 @@ import {studentProfile} from './fixtures.jsx';
 import SpecSugar from '../support/spec_sugar.jsx';
 import createSpyObj from '../support/createSpyObj.js';
 import {merge} from '../../../app/assets/javascripts/helpers/react_helpers.jsx';
+import RestrictedNotesPageContainer from '../../../app/assets/javascripts/restricted_notes/RestrictedNotesPageContainer.js';
 
 describe('RestrictedNotesPageContainer', function() {
   const ReactDOM = window.ReactDOM;
-  const RestrictedNotesPageContainer = window.shared.RestrictedNotesPageContainer;
 
   const helpers = {
     renderInto: function(el, props) {
